Update for recent Zig 0.12, use .NET 6.0
This commit updates the repo to build with more recent versions of Zig
0.12, namely updating for semantic changes, e.g. enforcing const when
values are not mutated, build API changes, generalization of readInt
and writeInt, etc. This also updates the zig-args and parser-toolkit
dependencies to accommodate.

Additionally, this updates svg2tvgt to build with .NET 6.0 versus 5.0,
which is EOL and has been removed out of nixpkgs. This built without any
other changes necessary.